formulae: move stretchng parameters to separate file
