+++ libmpdata++/formulae/fractal_reconstruction.hpp
@@ -15,77 +15,6 @@ namespace libmpdataxx
   {
     namespace fractal
     {
-      namespace stretching_parameters
-      {
-        // stretching parameters for velocity from DNS
-        // the function used is a fit to the E. Akinlabi data 
-        // the data is in the stretching_parameter_data subdirectory
-        namespace DNS_vel 
-        {
-          // CDF of stretching parameter d; assuming CDF = A d^B + C under condition that CDF(0.5)=0 and CDF(1)=1
-          template<class real_t>
-          static real_t CDF_of_d(const real_t &d)
-          {
-            const real_t B = -0.39091161; 
-            return (pow(d,B) - pow(0.5,B)) / (1. - pow(0.5,B));
-          }
-          // inverse function: d of a CDF
-          template<class real_t>
-          static real_t d_of_CDF(const real_t &CDF)
-          {
-            const real_t B = -0.39091161; 
-            return pow((1.-pow(0.5,B))*CDF + pow(0.5,B), 1./B);
-          }
-          template<class real_t>
-          struct d_of_CDF_fctr_DNS
-          {
-            real_t operator()(const real_t &CDF) const
-            {
-              return CDF < 0 ? -d_of_CDF(-CDF) : d_of_CDF(CDF);
-            }
-            BZ_DECLARE_FUNCTOR(d_of_CDF_fctr_DNS);
-          };
-        };
-        // stretching parameters for potential temperature in a LES of marine cumulus
-        namespace LES_th_rv
-        {
-          // available distributions of the stretching parameter
-          // TODO: move to libmpdata++ opts?
-          enum class d_distro {rv_subsaturated, rv_supersaturated, th_subsaturated, th_supersaturated};
-          
-          template<class real_t>
-          class d_of_CDF_fctr_LES
-          {
-            std::vector<std::pair<real_t, real_t>> d_CDF_vctr;
-
-            public:
-
-            d_of_CDF_fctr_LES(d_distro dd)
-            {
-              switch(dd)
-              {
-                case th_subsaturated:
-                  d_CDF_th_subsaturated(d_CDF_vctr);
-                  break;
-                default:
-                  std::runtime_error("libmpdata++: invalid d_distro type");
-              }
-            };
-
-            real_t operator()(const real_t &CDF) const
-            {
-              auto pos = std::lower_bound(d_CDF_vctr.begin(), d_CDF_vctr.end(), CDF, 
-                [](const std::pair<real_t, real_t> &pair, real_t val)
-                {
-                  return pair.second < val;
-                });
-              return pos->first;
-            }
-            BZ_DECLARE_FUNCTOR(d_of_CDF_fctr_LES);
-          };
-        };
-      };
-
       // (tri-)linear interpolation
       // 3d version
       template<int d, class real_t, class arr_t, class rng_t>
